+++ crates/zng-ext-font/src/shaping.rs
@@ -797,7 +797,9 @@ impl ShapedText {
 
                 let first_line = self.lines.first_mut();
                 first_line.x_offset = first.origin.x.0 as f32;
-                first_line.width = first.size.width.0 as f32;
+                // width is the same measured, unless the parent inliner changed it to fill,
+                // in that case we need the original width in `reshape_lines_justify`.
+                // first_line.width = first.size.width.0 as f32;
             }
             if !first_segs.is_empty() {
                 // parent set first_segs.
@@ -835,7 +837,8 @@ impl ShapedText {
 
                 let last_line = self.lines.last_mut();
                 last_line.x_offset = last.origin.x.0 as f32;
-                last_line.width = last.size.width.0 as f32;
+                // width is the same measured, unless the parent inliner changed it to justify, that is handled later.
+                // last_line.width = last.size.width.0 as f32;
             }
             if !last_segs.is_empty() {
                 // parent set last_segs.
@@ -996,16 +999,13 @@ impl ShapedText {
     }
 
     fn justify_lines_range(&self) -> ops::Range<usize> {
-        let mut range = 0..self.lines_len().saturating_sub(1); // skip last line
+        let mut range = 0..self.lines_len();
 
-        if self.is_inlined {
-            // justify on inline not implemented
-            range.start += 1;
+        if !self.is_inlined {
+            // skip last line
             range.end = range.end.saturating_sub(1);
-            if range.start > range.end {
-                range = 0..0
-            }
         }
+        // else inlined fills the first and last line rects
 
         range
     }
@@ -1026,6 +1026,7 @@ impl ShapedText {
         let range = self.justify_lines_range();
 
         let fill_width = self.align_size.width.0 as f32;
+        let last_li = range.end.saturating_sub(1);
 
         for li in range.clone() {
             let mut count;
@@ -1033,6 +1034,16 @@ impl ShapedText {
             let mut line_seg_range;
             let mut offset = 0.0;
             let mut last_is_space = false;
+
+            let mut fill_width = fill_width;
+            if self.is_inlined {
+                // inlining parent provides the fill space for the first and last segment
+                if li == 0 {
+                    fill_width = self.first_line.width().0 as f32;
+                } else if li == last_li {
+                    fill_width = self.last_line.width().0 as f32;
+                }
+            }
 
             {
                 // line scope